feat(locale): seperate locale core into common dictionary
