+++ src/modules/action-sheet/index.js
@@ -1,6 +1,6 @@
 import ActionSheet from '../../components/action-sheet/action-sheet.vue'
 import addActionSheet from './api'
-import Locale from '../locale'
+import Locale from '../../common/locale'
 
 ActionSheet.install = function (Vue) {
   Vue.component(ActionSheet.name, ActionSheet)
