+++ src/modules/cascade-picker/index.js
@@ -2,7 +2,7 @@ import Picker from '../../components/picker/picker.vue'
 import CascadePicker from '../../components/cascade-picker/cascade-picker.vue'
 import addCascadePicker from './api'
 import addPicker from '../picker/api'
-import Locale from '../locale'
+import Locale from '../../common/locale'
 
 CascadePicker.install = function (Vue) {
   Vue.component(Picker.name, Picker)
