+++ src/modules/locale/index.js
@@ -1,53 +1,3 @@
-import defaultMessages from '../../locale/lang/zh-CN'
-import { warn } from '../../common/helpers/debug'
-import { isUndef, isNumber } from '../../common/helpers/util'
-import { formatDate } from '../../common/lang/date'
+import Locale from '../../common/locale'
 
-let Vue
-let proto
-
-const DEFAULT_LANG = 'zh-CN'
-
-const locale = {
-  name: 'locale',
-  install (_Vue) {
-    if (locale.installed) return
-    Vue = _Vue
-    proto = Vue.prototype
-    Vue.util.defineReactive(proto, '$cubeLang', DEFAULT_LANG)
-    proto['$cubeMessages'] = { [DEFAULT_LANG]: defaultMessages }
-    locale.installed = true
-  },
-  use (lang, messages) {
-    proto['$cubeLang'] = lang
-    const cubeMessages = proto['$cubeMessages']
-    // if messages have never been stored in vue.prototye
-    if (!(lang in cubeMessages)) {
-      cubeMessages[[lang]] = messages
-    }
-  },
-  helpers: {
-    toLocaleDateString (config, formatRules) {
-      /**
-       * Safari don't support formatRules like
-       * 'yyyy-MM-dd hh:mm:ss', so transfer it to 'yyyy/MM/dd hh:mm:ss'
-       */
-      const compatibleConfig = isNumber(config) ? config : config.replace(/-/g, '/')
-      const date = new Date(compatibleConfig)
-      /* istanbul ignore if */
-      if (isUndef(formatRules)) return date.toDateString()
-      return formatDate(date, formatRules)
-    }
-  },
-  addHelper (fnName, fn) {
-    // check existed helper fn
-    /* istanbul ignore if */
-    if (fnName in locale.helpers) {
-      warn(`${fnName} has already been registered on helpers function, please change another name`)
-      return
-    }
-    locale.helpers[fnName] = fn
-  }
-}
-
-export default locale
+export default Locale
