+++ src/modules/select/index.js
@@ -1,7 +1,7 @@
 import Picker from '../../components/picker/picker.vue'
 import Select from '../../components/select/select.vue'
 import addPicker from '../picker/api'
-import Locale from '../locale'
+import Locale from '../../common/locale'
 
 Select.install = function (Vue) {
   Vue.component(Picker.name, Picker)
