+++ src/modules/time-picker/index.js
@@ -2,7 +2,7 @@ import Picker from '../../components/picker/picker.vue'
 import TimePicker from '../../components/time-picker/time-picker.vue'
 import addTimePicker from './api'
 import addPicker from '../picker/api'
-import Locale from '../locale'
+import Locale from '../../common/locale'
 
 TimePicker.install = function (Vue) {
   Vue.component(Picker.name, Picker)
